fix: refactor of one of the folders
