+++ packages/network/src/provider/utils/const/blocks/blocks.ts
@@ -1,5 +1,5 @@
 import { Hex } from '@vechain/sdk-core';
-import { type BlockQuantityInputRPC } from '../../rpc-mapper';
+import { type BlockQuantityInputRPC } from '../../rpc-mapper/types';
 
 /**
  * Get the correct block number for the given block number.
